+++ packages/carp_survey_package/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 1.8.3
+## 1.8.4
 
 * added a `SurveyStatus` to the survey results to know if a survey was "submitted" or "canceled"
+* refix of issue [#342](https://github.com/cph-cachet/carp_studies_app/issues/342)
 
 ## 1.8.2